robust handling of constants; comments; use encrypt_message function; move get-request-from-proposal to proto utils
